+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 * Adds `-useManifests` option to generate applets and workflows whose inputs and outputs are manifest files
 * Fixes issue with using both streaming and non-streaming file inputs in the same task
+* Fixes issue with scatter as the first element of a workflow
 
 ## 2.3.1 03-03-2021